% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_exposure_table.R
\name{create_exposure_table}
\alias{create_exposure_table}
\title{A function to generate a time-varying variable for simulation}
\usage{
create_exposure_table(ppt_tbl, slope = -9e-04, slope_sd = 1e-04, n_years = 2L)
}
\arguments{
\item{ppt_tbl}{A data.table; the return value of create_ppt_table}

\item{slope}{the average slope for the linear temporal trend}

\item{slope_sd}{the standard deviation of the (normally distributed) distribution of participant-specific slopes
(which is centered around slope)}

\item{n_years}{Number of years over raw spline data to take average}
}
\description{
A function to generate participant-specific random variables that vary smoothly over time.
}
\details{
First the function generates four temporal trends:
two temporal trends are random curves (splines fit to random points).
There is also a constant term and a linear term.

Each participant has their own coefficient to each temporal trend.
the slope argument determines the slope of the linear trend
}
