% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hazard_table.R
\name{create_hazard_table}
\alias{create_hazard_table}
\title{Generate a baseline hazard for simulation}
\usage{
create_hazard_table(ppt_tbl, avg_daily_hazard = 3e-06, shape = 7.5)
}
\arguments{
\item{ppt_tbl}{A data.table; the return value of create_ppt_table}

\item{avg_daily_hazard}{A length-1 numeric vector signifying the desired average hazard
(ie probability of an event conditional on being at risk at that time)
return a vector the of length n. Signifies the age at enrollment.}

\item{shape}{Controls the variabililty around avg_daily_hazard
return a vector the of length n. Signifies calendar time at study entry (ie enrollment date).}
}
\value{
a data.table with ppt_id, time, numeric column hazard,  integer columns age_tv (in units of days) age_since65_tv
(both in units of days since 365.25*65)
}
\description{
A function to randomly generate baseline hazard as a function of time for each ppt
}
\details{
The return table has a row for each combination of ppt and time,
but the hazard is the same for all ppts.
}
