% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ppt_table.R
\name{create_ppt_table}
\alias{create_ppt_table}
\title{Generate time-varying exposure data for simulation}
\usage{
create_ppt_table(
  n,
  baseline_age_expr = as.integer(floor(55 * 365.25 + rbinom(n, size = 30, prob = 0.3) *
    365.25)),
  t_start = pmin(as.integer(floor(as.numeric(as.IDate("2000-01-01") + abs(rnorm(n, 300,
    sd = 400)))), as.integer(as.IDate("2017-01-01")))),
  t_exit = as.integer(as.IDate("2017-01-01"))
)
}
\arguments{
\item{n}{The desired number of participants (ie rows) of the output}

\item{baseline_age_expr}{An expression written as a function of n which will
return a vector the of length n. Signifies the age at enrollment.}

\item{t_start}{An expression written as a function of n which will
return a vector the of length n. Signifies calendar time at study entry (ie enrollment date).}

\item{t_exit}{An expression written as a function of n which will
return a vector the of length n. Signifies calendar time for last possible date of follow-up.}
}
\value{
A data.table with columns ppt_id, time, pm25, and pm25_avg (where pm25_avg is just the
rolling average of pm25 where the size of the rolling window is defined by n_years)
}
\description{
A function to generate participant-specific random variables that vary smoothly over time
}
\details{
Defaults generate baseline ages as integers, randomly, with mean of ~64*365.25.
Default enrollment date is generated between 2001-01-01 and 2017-01-01.
Default t_exit (last possible day of enrollment) is 2017-01-01.
}
\examples{
create_ppt_table(10,
                baseline_age=as.integer(floor(55*365.25 + rbinom(n,size=30, prob=.3)*365.25)),
                t_start=pmin(
                  as.integer(floor(as.numeric(
                   as.IDate("2000-01-01")+
                      abs(rnorm(n,300,sd=400))
                  )),
                  as.integer(as.IDate("2017-01-01"))
                  )),
                t_exit=as.integer(as.IDate("2017-01-01"))
)

}
