% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cox_results.R
\name{extract_coef_table}
\alias{extract_coef_table}
\title{extract coefficients from a cox model and turn them into a table}
\usage{
extract_coef_table(model, RHS = NULL)
}
\arguments{
\item{model}{A model returned by coxph}

\item{RHS}{description of the model here. if NULL this will try to extrac the formulat from the call
but this may not work if you're fitting the model programmatically (ie via a loop)}
}
\value{
a data.table of coefficients from the models including a column "RHS" for the
RHS of the formula from the call
}
\description{
take a single cox model and extract the coeffients from summary
}
