% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_exposure_matrix.R
\name{generate_exposure_matrix}
\alias{generate_exposure_matrix}
\title{Generate a table containing time-varying exposure for use with simulate_survival}
\usage{
generate_exposure_matrix(
  x,
  ppt_id,
  mean = 15,
  slope = -9e-04,
  slope_sd = 1e-04,
  random_spline_points_range = 25,
  degree = 10,
  proportion_points_sampled = 0.1,
  dofoverlaps = TRUE
)
}
\arguments{
\item{x}{A vector of time points}

\item{ppt_id}{A vector of ppt_ids}

\item{mean}{The expected value of the curve at the midpoint of time.}

\item{slope}{the average slope of centered time.}

\item{slope_sd}{the standard deviation of the (normally distributed) distribution of participant-specific slopes
(which is centered around slope)}

\item{random_spline_points_range}{Passed to runif to generate points to create random splines.
not the actual range of the data around the mean because the spline can go beyond this range
and becasue these splines are multiplied by coefficients, but altering this will alter the
range of the variabilty around the linear trend determined by slope and mean.}

\item{degree}{passed to generate_random_spline which is passed to bs}

\item{proportion_points_sampled}{determines number of points sampled to make the curve}

\item{dofoverlaps}{if TRUE, final dataset gets subset to ppt-specific follow-up periods,
otherwise the dataset is left at every combination of ppt and time whether or not the ppt
was being followed at that time (including times before and after the start of follow-up that
generated so that the actual used portion of the curve isn't on the edge which is variable.}
}
\value{
a matrix with length(x) rows and length(ppt_id) columns
}
\description{
A function to generate a table containing a participant-specific
time-dependent covariate that varies smoothly over time.
the time-dependent covariates differ between individuals but are correlated because
they are generated from a random linear combination of the same underlying splines.
}
\details{
All combinations of x and ppt_id arguments are crossed to create a matrix.
}
\examples{

ppt_tbl <- generate_ppt_table(1000)
e <- expand_ppt_table(ppt_tbl)

setkey(e, ppt_id, time)

times <- e$time[e$ppt_id==1]
ppts <- unique(e$ppt_id)
exposures <- generate_exposure_matrix(times,ppts)
e[, exposure:=as.vector(exposures)]

#plot a subset of participant time-series
matplot(times,exposures[,50:70] ,col=1:ncol(exposures ),type="l",lty=1)

}
