% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hazard_table.R
\name{generate_hazard_table}
\alias{generate_hazard_table}
\title{Generate a table containing baseline hazard for use with simulate_survival}
\usage{
generate_hazard_table(
  ppt_tbl,
  daily_hazard_mean = 3e-06,
  daily_hazard_range = 1e-06,
  daily_hazard_slope = -1e-10,
  degree = 20,
  proportion_points_sampled = 0.1
)
}
\arguments{
\item{ppt_tbl}{A data.table; the return value of create_ppt_table}

\item{daily_hazard_mean}{A length-1 numeric vector signifying the desired average hazard
(ie probability of an event conditional on being at risk at that time)
Since the slope (specified in \code{daily_hazard_slope}) is on centered time, the expected
value of the curve at the midpoint of the input times equals daily_hazard_mean.}

\item{daily_hazard_range}{parameters passed to runif to determine range of possible points used
to generate curve. not the actual possible range since the spline curve may exceed the range of the data.}

\item{daily_hazard_slope}{The slope of centered time on hazard curve (specify in units of hazard/day).}

\item{degree}{passed to generate_random_spline which is passed to bs}

\item{proportion_points_sampled}{determines number of points sampled to make the random spline curve}
}
\value{
a data.table with ppt_id, time (integer days since 1970-01-01),
numeric column baseline_hazard,
integer columns age_tv and age_since65_tv
(in units of days since birthdate and days since age 65 respectively)
}
\description{
The return table has a row for each combination of ppt and time,
but the hazard is the same for all ppts.
}
\details{
Degree affects the flexibility of the curve but the flexiblity of the curve
is limited by the number of points that are sampled.
proportion_points_sampled
}
\examples{
ppt_tbl <- generate_ppt_table(1000)
hazard_tbl <- generate_hazard_table(ppt_tbl)
hazard_tbl[ppt_id==1, plot(time,baseline_hazard, type="l")]
}
