% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_curve.R
\name{generate_random_spline}
\alias{generate_random_spline}
\title{Create a random curve using bs splines}
\usage{
generate_random_spline(x_range, n_points, random_points_expr, degree)
}
\arguments{
\item{x_range}{A length-2 numeric vector specifying the inclusive interval over which the
curve should be defined.}

\item{n_points}{For the random splines, the number of points (not including end points) that will be
used to generate the spline.}

\item{random_points_expr}{For the random splines, an expression of n used to generate
the y value of the random points. See examples}

\item{degree}{The degree for the random splines. Passed to bs function.}
}
\value{
a function of x
}
\description{
Generate a random set of x,y points then fit a spline through them.
}
\examples{
set.seed(20)
f <- generate_random_spline(c(1,100),10, stats::rnorm(n),degree=4)
plot(1:100, f(1:100),type="l")
}
