% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_events.R
\name{melt_events}
\alias{melt_events}
\title{Melt multiple type columns to long format
Reshape to long a data.table with survival data for multiple event types.}
\usage{
melt_events(x, id, measure.vars, value.name, type)
}
\arguments{
\item{x}{A data.table}

\item{id}{A length-1 character vector corresponding to a column of x that is a unique identifier.}

\item{measure.vars}{A list where each element is a length-n character vector corresponding to column names in x.
The character vectors must be in the same order.  Each element of the list represents a set of
columns that will be turned into a single column in the return.
(such that measure.vars[\link{1}]\link{1} corresponds to measure.vars[\link{2}]\link{1},measure.vars[\link{1}]\link{2}
corresponds to measure.vars[\link{2}]\link{2},etc).
Typically, one element of measure.vars is a character vector corresponding to column names of follow-up dates/days and
and another  element of measure.vars is a character vector corresponding to column names of binary event indicators.}

\item{value.name}{A character vector the length of measure.vars designating desired the column names
in the return. Each element of value.name corresponds to the column that will be created from
each element of measure.vars}

\item{type}{A vector the length of each element of measure.vars denoting the type of event.
ie, type\link{1} denotes the type of event of measure.vars[\link{1}]\link{1},measure.vars[\link{2}]\link{1}, etc and
type\link{2} denotes the type of event of measure.vars[\link{1}]\link{2},measure.vars[\link{2}]\link{2}, etc.
type will become an indicator column in the return}
}
\value{
If there's only a single baseline date for all event types, ignore it for the purpose of this function
then merge it back in to the result. The same goes for any other variables in x that you want to be in y
that don't need to be melted.
}
\description{
Melt multiple type columns to long format
Reshape to long a data.table with survival data for multiple event types.
}
\examples{
x <- data.table(id=c(1,2,3),baseline_date=c(10,5,3),
mi_fu_date=c(15,8,20),mi_indicator=c(1,1,0),death_fu_date=c(26,8,20),death_indicator=c(0,1,0))
x_long <- melt_events(x,id="id",
 measure.vars=list(c("mi_fu_date","death_fu_date"),c("mi_indicator","death_indicator")),
 value.name=c("fu_date","event_indicator"),
 type=c("mi","death")
)
x_long <- x[,list(id,baseline_date)][x_long,on="id"] #merge baselin_date back in

}
