% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep}
\alias{prep}
\title{prep function}
\usage{
prep(x)
}
\arguments{
\item{x}{is a data.table with the following columns:
\itemize{
\item id: a unique identifier for participant
\item age: age (in years, rounded down ) at baseline.
\item time: days since baseline
\item event: 0/1
}}
}
\description{
prep function
}
