% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_ppt_table.R
\name{prep_tv_dataset}
\alias{prep_tv_dataset}
\title{expand participant table, add hazard, and exposures}
\usage{
prep_tv_dataset(ppt_tbl, include_exposure = TRUE)
}
\arguments{
\item{ppt_tbl}{typically the return of generate_ppt_tbl, possibly with additional columns added.
specifically: a data.table with columns ppt_id (assummed to be unique), t_start, t_exit, baseline_age,
and possibly other columns}

\item{include_exposure}{if TRUE, adds an exposure column. Turn this off if you're not using the exposure column since
it takes time to generate this}
}
\value{
a data.table where the number of rows will be the number of participants times the total number of unique integer
time points in the set union of all the [t_start,t_exit] intervals in ppt_tbl.
Ever ppt_id receives a row for every time point, whether or not the participant was at risk at that time.
it has the following columns:
\itemize{
\item ppt_id
\item time
\item at_risk (indicator for whether participant was at risk during this time. this is just
whether this time is in that participant's interval which for the purpose of the risk
is considered open on the left and closed on the right.
ie this is an indicator for whether time is in (t_start,t_exit].
\item age_tv time-varying age. calculated for convenience
\item age_since65_tv time-varying age in units of days since age 65. assumes baseline_age was in units of days
\item exposure
\item baseline_hazard
\item any other columns that were in ppt_tbl
}
}
\description{
calls expand_ppt_table to expand a participant table, then calls
generate_exposure_matrix and generate_baseline_hazard to add those columns.
}
\details{
If you need to use non-default options to generate_exposure_matrix or generate_baseline_hazard,
this function could easily be edited to pass arguments to those respective functions as a list.
}
