% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_survival2.R
\name{simulate_survival}
\alias{simulate_survival}
\title{this is a title}
\usage{
simulate_survival(
  x,
  covar_names = NULL,
  covar_coefs = NULL,
  return_hazard = FALSE
)
}
\arguments{
\item{x}{a data.table containing at a mininum a unique identifier column named ppt_id, an integer time column
named time, a positive numeric column named baseline_hazard. Participants are assumed to be at risk if a time/participant combination exists as a row.
Within groups defined by participants, the values of time (or number of rows in each group) need not be the same.
May also contain covariates.}

\item{covar_names}{a vector of column names in x corresponding to numeric covariates}

\item{covar_coefs}{A vector of coefficients (same length as covar_names)}

\item{return_hazard}{If TRUE, the return data.table will contain an attribute which itself is a data.table
containing ppt-specific hazard at each time-point.}
}
\value{
a data.table
}
\description{
this is a summary
}
\details{
note that the t_exit column in x represents the last day on which a ppt could be observed. typically
the end of follow-up, and may be the same day for everyone.
}
