% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cox_results.R
\name{summarize_simulation_table}
\alias{summarize_simulation_table}
\title{calculate average coefficients from a table}
\usage{
summarize_simulation_table(results_table)
}
\arguments{
\item{results_table}{Tables of coefficients. Use rbindlist on a list of results from extract_coef_table}
}
\value{
a data.table of summarized coefficients
}
\description{
Calculate average coefficients from a table containing results from several model runs
}
